requirements.txt
openai
langchain
GitPython
chromadb
numpy==1.21.6